pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "vllm-monitor"
version = "0.1.0"
description = "Live terminal dashboard for monitoring vLLM inference services and GPU stats"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "subterra" },
]
keywords = ["vllm", "monitor", "dashboard", "gpu", "inference"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console :: Curses",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: System :: Monitoring",
]

[project.optional-dependencies]
yaml = ["pyyaml>=6.0"]

[project.scripts]
vllm-monitor = "vllm_monitor.cli:main"

[project.urls]
Homepage = "https://github.com/subterra/vllm-monitor"
Issues = "https://github.com/subterra/vllm-monitor/issues"
